+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Implemented `WeatherMapper` to convert `WeatherOpenMap` to `Weather` domain model.
 - Added `WeatherMapperTest` to verify mapping logic.
 - Added missing file `application.properties` in test folder
+- Added method `getWeather` to class `OpenWeatherMapService` and relative test.
 
 ### Removed